requirements.txt
requests==2.31.0
python-dotenv==1.0.0
spacy==3.7.2
pytest==8.0.0
pytest-cov==4.1.0
transformers
torch


